+++ CHANGELOG.md
@@ -1,11 +1,16 @@
 # Changelog
 
+## Major.Minor.Revision (Month Day, Year)
+
+### BUG FIXES
+
+* Fix bug dealing with multiple profiles in okta.yaml [#187](https://github.com/okta/okta-aws-cli/pull/187), thanks [@MatthewJohn](https://github.com/MatthewJohn)!
+
 ## 2.1.0 (February 15, 2024)
 
 ### ENHANCEMENTS
 
-* Multiple okta-aws-cli configurations in `okta.yaml` by AWS profile name.
- [#162](https://github.com/okta/okta-aws-cli/pull/162), thanks [@MatthewJohn](https://github.com/MatthewJohn)!
+* Multiple okta-aws-cli configurations in `okta.yaml` by AWS profile name.  [#162](https://github.com/okta/okta-aws-cli/pull/162), thanks [@MatthewJohn](https://github.com/MatthewJohn)!
 
 * Explicitly set AWS Region with CLI flag `--aws-region` [#174](https://github.com/okta/okta-aws-cli/pull/174), thanks [@euchen-circle](https://github.com/euchen-circle), [@igaskin](https://github.com/igaskin)!